requirements.txt
numpy==1.23.5
pandas==1.5.3
tabula-py==2.9.0
JPype1==1.4.1
openpyxl==3.1.2
python-dateutil==2.8.2